+++ packages/components/accordion/package.json
@@ -43,7 +43,6 @@
     "@oku-ui/primitive": "latest",
     "@oku-ui/provide": "latest",
     "@oku-ui/roving-focus": "latest",
-    "@oku-ui/slot": "latest",
     "@oku-ui/use-composable": "latest",
     "@oku-ui/utils": "latest",
     "@types/node": "^18.17.12",
