+++ packages/components/accordion/src/accordionImplMultiple.ts
@@ -1,55 +1,12 @@
-import { type PropType, computed, defineComponent, h, mergeProps, reactive, toRefs } from 'vue'
+import { computed, defineComponent, h, mergeProps, reactive, toRefs } from 'vue'
 import { primitiveProps } from '@oku-ui/primitive'
-import type { OkuElement } from '@oku-ui/primitive'
-import { useForwardRef } from '@oku-ui/use-composable'
-import { useControllable } from '../../../core/use-composable/dist'
-import { AccordionCollapsibleProvider, AccordionValueProvider, scopeAccordionProps } from './utils'
-import { OkuAccordionImpl, accordionImplProps } from './accordionImpl'
-import type { AccordionImplEmits, AccordionImplProps } from './accordionImpl'
+import { reactiveOmit, useControllable, useForwardRef } from '@oku-ui/use-composable'
+import {
+  ACCORDION_IMPL_MULTIPLE_NAME, AccordionCollapsibleProvider, type AccordionImplMultipleNativeElement, AccordionValueProvider, accordionImplMultipleProps,
+  scopeAccordionProps,
+} from './props'
+import { OkuAccordionImpl } from './accordionImpl'
 
-const ACCORDION_IMPL_MULTIPLE_NAME = 'OkuAccordionImplMultiple'
-export type AccordionImplMultipleNativeElement = OkuElement<'div'>
-
-export type AccordionImplMultipleElement = HTMLDivElement
-
-export interface AccordionImplMultipleProps extends AccordionImplProps {
-  /**
-   * The controlled stateful value of the accordion item whose content is expanded.
-   */
-  value?: string[]
-  /**
-   * The value of the item whose content is expanded when the accordion is initially rendered. Use
-   * `defaultValue` if you do not need to control the state of an accordion.
-   */
-  defaultValue?: string[]
-}
-export interface AccordionImplMultipleEmits extends AccordionImplEmits {
-  /**
-   * The callback that fires when the state of the accordion changes.
-   */
-  valueChange: [value: string[]]
-}
-export const accordionImplMultipleProps = {
-  props: {
-    ...accordionImplProps.props,
-    modelValue: {
-      type: [Array, undefined] as PropType<string[] | undefined>,
-      default: undefined,
-    },
-    defaultValue: {
-      type: [Array, undefined] as PropType<string[] | undefined>,
-      default: undefined,
-    },
-  },
-  emits: {
-    ...accordionImplProps.emits,
-    /**
-   * The callback that fires when the state of the accordion changes.
-   */
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    valueChange: (value: string[]) => true,
-  },
-}
 const accordionImplMultiple = defineComponent({
   name: ACCORDION_IMPL_MULTIPLE_NAME,
   inheritAttrs: false,
@@ -61,11 +18,10 @@ const accordionImplMultiple = defineComponent({
   emits: accordionImplMultipleProps.emits,
   setup(props, { slots, emit, attrs }) {
     const {
-      modelValue: valueProp,
+      value: valueProp,
       defaultValue,
       ...accordionMultipleProps
     } = toRefs(props)
-
     const { state, updateValue } = useControllable({
       prop: computed(() => valueProp.value),
       defaultProp: computed(() => defaultValue.value),
@@ -74,25 +30,27 @@ const accordionImplMultiple = defineComponent({
       },
       initialValue: undefined,
     })
-    const handleItemOpen = (itemValue: string) => {
-      state.value.push(itemValue)
-      updateValue(state.value)
-    }
-    const handleItemClose = (itemValue: string) => {
-      const index = state.value.indexOf(itemValue)
-      state.value.splice(index, 1)
-      updateValue(state.value)
-    }
 
     const forwardRef = useForwardRef()
 
-    const _accordionMultipleProps = reactive(accordionMultipleProps)
+    const _reactive = reactive(accordionMultipleProps)
+    const _accordionMultipleProps = reactiveOmit(_reactive, (key, _value) => key === undefined)
 
     AccordionValueProvider({
       scope: props.scopeOkuAccordion,
-      modelValue: computed(() => state.value),
-      onItemOpen: handleItemOpen,
-      onItemClose: handleItemClose,
+      value: computed(() => state.value),
+      onItemOpen: (e) => {
+        if (!state.value)
+          updateValue([])
+
+        state.value.push(e)
+        updateValue(state.value)
+      },
+      onItemClose: (e) => {
+        const index = state.value.indexOf(e)
+        state.value.splice(index, 1)
+        updateValue(state.value)
+      },
     })
     AccordionCollapsibleProvider({
       scope: props.scopeOkuAccordion,
