+++ packages/components/accordion/src/utils.ts
@@ -1,67 +1,3 @@
-import type { Scope } from '@oku-ui/provide'
-import { ScopePropObject, createProvideScope } from '@oku-ui/provide'
-import { createCollection } from '@oku-ui/collection'
-import { createCollapsibleScope } from '@oku-ui/collapsible'
-import type { Ref } from 'vue'
-import type { AccordionTriggerElement } from './accordionTrigger'
-import type { AccordionImplProps } from './accordionImpl'
-
-export const ACCORDION_KEYS = ['Home', 'End', 'ArrowDown', 'ArrowUp', 'ArrowLeft', 'ArrowRight']
-
-export type Direction = 'ltr' | 'rtl'
-export type SelectionType = 'single' | 'multiple'
-
-export const ACCORDION_NAME = 'OkuAccordion'
-export const ITEM_NAME = 'OkuAccordionItem'
-
-export const { CollectionItemSlot, CollectionProvider, CollectionSlot, useCollection, createCollectionScope } = createCollection<AccordionTriggerElement>(ACCORDION_NAME)
-
-export type ScopeAccordion<T> = T & { scopeOkuAccordion?: Scope }
-
-export const scopeAccordionProps = {
-  scopeOkuAccordion: {
-    ...ScopePropObject,
-  },
-}
-
-export const [createAccordionProvider, createAccordionScope] = createProvideScope(ACCORDION_NAME, [
-  createCollectionScope,
-  createCollapsibleScope,
-])
-
-export const useCollapsibleScope = createCollapsibleScope()
-
-type AccordionValueProviderValue = {
-  modelValue: Ref<string[] | string>
-  onItemOpen(value: string): void
-  onItemClose(value: string): void
-}
-
-export const [AccordionValueProvider, useAccordionValueInject]
-  = createAccordionProvider<AccordionValueProviderValue>(ACCORDION_NAME)
-
-export const [AccordionCollapsibleProvider, useAccordionCollapsibleInject] = createAccordionProvider(
-  ACCORDION_NAME,
-  { collapsible: false },
-)
-
-type AccordionImplContextValue = {
-  disabled?: Ref<boolean | undefined>
-  direction: Ref<AccordionImplProps['dir']>
-  orientation: Ref<AccordionImplProps['orientation']>
-}
-export const [AccordionImplProvider, useAccordionInject] = createAccordionProvider<AccordionImplContextValue>(
-  ACCORDION_NAME,
-)
-
 export function getState(open?: boolean) {
   return open ? 'open' : 'closed'
 }
-
-type AccordionItemContextValue =
-{ open?: Ref<boolean>
-  disabled?: Ref<boolean | undefined>
-  triggerId: Ref<string>
-}
-export const [AccordionItemProvider, useAccordionItemInject]
-  = createAccordionProvider<AccordionItemContextValue>(ITEM_NAME)
